+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Added "sort keys" toggle for objects in `struct` view
 - Added "sort keys" and "dict mode" actions for objects in `signature` view
 - Extended `menu-item` to take data as item text when it's a string
-- Changed `block` view to pass `onInit` and `onChange` handlers if any to its content
+- Extended `block` view to pass `onInit` and `onChange` handlers if any to its content
+- Extended `checkbox`, `input`, `select` and `tabs` views to pass `data` and `context` to `onInit` and `onChange` handler beside value and name
 - Changed `utils.createElement()` to ignore attributes with not own keys or `undefined` value
 - Fixed edge cases for value stat output in `signature` view
 - Fixed scrolling issue for `tabs` view when used in sidebar